clippy

+++ drv/cosmo-hf/src/hf.rs
@@ -859,13 +859,9 @@ impl idl::InOrderHostFlashImpl for ServerImpl {
         &mut self,
         _: &RecvMessage,
     ) -> Result<(), RequestError<HfError>> {
-        let base = Self::flash_addr_for(SECTOR_SIZE_BYTES as u32, self.dev)
-            .unwrap_lite();
-        let hash = self.drv.sha3(
-            base,
-            SECTOR_SIZE_BYTES as u32,
-            SLOT_SIZE_BYTES as u32,
-        )?;
+        let base =
+            Self::flash_addr_for(SECTOR_SIZE_BYTES, self.dev).unwrap_lite();
+        let hash = self.drv.sha3(base, SECTOR_SIZE_BYTES, SLOT_SIZE_BYTES)?;
 
         self.measurement.copy_from_slice(&hash);
         Ok(())
